+++ apps/pie-storybook/stories/pie-chip.stories.ts
@@ -1,36 +1,112 @@
-import { html } from 'lit';
+import { html, nothing } from 'lit';
 
 /* eslint-disable import/no-duplicates */
 import '@justeattakeaway/pie-chip';
-import { ChipProps } from '@justeattakeaway/pie-chip';
+import { ChipProps as ChipPropsBase, variants } from '@justeattakeaway/pie-chip';
 /* eslint-enable import/no-duplicates */
+import '@justeattakeaway/pie-icons-webc/IconHeartFilled';
 
-import { type StoryMeta } from '../types';
-import { createStory } from '../utilities';
+import type { StoryMeta, SlottedComponentProps } from '../types';
+import { createStory, type TemplateFunction, sanitizeAndRenderHTML } from '../utilities';
 
+type ChipProps = SlottedComponentProps<ChipPropsBase> & { showIcon: boolean };
 type ChipStoryMeta = StoryMeta<ChipProps>;
 
-const defaultArgs: ChipProps = {};
+const defaultArgs: ChipProps = {
+    variant: 'default',
+    disabled: false,
+    isSelected: false,
+    isLoading: false,
+    isDismissible: false,
+    showIcon: false,
+    slot: 'String',
+};
 
 const chipStoryMeta: ChipStoryMeta = {
     title: 'Chip',
     component: 'pie-chip',
-    argTypes: {},
+    argTypes: {
+        variant: {
+            description: 'Set the variant of the chip.',
+            control: 'select',
+            options: variants,
+            defaultValue: {
+                summary: 'default',
+            },
+        },
+        disabled: {
+            description: 'If `true`, disables the chip.',
+            control: 'boolean',
+            defaultValue: {
+                summary: false,
+            },
+        },
+        isSelected: {
+            description: 'If `true`, the chip element will apply the selected styles.',
+            control: 'boolean',
+            defaultValue: {
+                summary: false,
+            },
+        },
+        isLoading: {
+            description: 'If `true`, displays a loading indicator inside the chip.',
+            control: 'boolean',
+            defaultValue: {
+                summary: false,
+            },
+        },
+        isDismissible: {
+            description: 'If `true`, displays a close icon to dismiss the chip component. <br /><br /> Can be only used if `isSelected` is set to true',
+            control: 'boolean',
+            defaultValue: {
+                summary: false,
+            },
+            if: { arg: 'isSelected', eq: true },
+        },
+        showIcon: {
+            description: 'Enable to see the example of Chip with icon.',
+            control: 'boolean',
+            defaultValue: {
+                summary: false,
+            },
+        },
+        slot: {
+            description: 'Content to place within the chip',
+            control: 'text',
+        },
+    },
     args: defaultArgs,
     parameters: {
         design: {
             type: 'figma',
-            url: '',
+            url: 'https://www.figma.com/file/OOgnT2oNMdGFytj5AanYvt/%5BCore%5D-Web-Component-Documentation-%5BPIE-3%5D?type=design&node-id=575-6726&mode=design&t=axoB6cYrus8n0tdC-0',
         },
     },
 };
 
 export default chipStoryMeta;
 
-// TODO: remove the eslint-disable rule when props are added
-// eslint-disable-next-line no-empty-pattern
-const Template = ({}: ChipProps) => html`
-    <pie-chip></pie-chip>
-`;
+const Template: TemplateFunction<ChipProps> = ({
+    variant,
+    disabled,
+    isSelected,
+    isLoading,
+    isDismissible,
+    showIcon,
+    slot,
+}) => html`
+           <pie-chip
+                variant="${variant}"
+                ?disabled="${disabled}"
+                ?isSelected="${isSelected}"
+                ?isLoading="${isLoading}"
+                ?isDismissible="${isDismissible}">
+                    ${showIcon ? html`<icon-heart-filled slot="icon"></icon-heart-filled>` : nothing}
+                    ${sanitizeAndRenderHTML(slot)}
+           </pie-chip>`;
 
-export const Default = createStory<ChipProps>(Template, defaultArgs)();
+const createCardStory = createStory<ChipProps>(Template, defaultArgs);
+
+export const Default = createCardStory();
+export const Outline = createCardStory({ variant: 'outline' });
+export const Ghost = createCardStory({ variant: 'ghost' });